refactor: Combine endpoints for downloadable resources
[#OCD-3601]
