+++ chpl/chpl-api/src/main/java/gov/healthit/chpl/web/controller/SearchViewController.java
@@ -1,14 +1,9 @@
 package gov.healthit.chpl.web.controller;
 
-import java.io.File;
-import java.io.IOException;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.HashSet;
 import java.util.Set;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.ff4j.FF4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,7 +20,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import gov.healthit.chpl.FeatureList;
 import gov.healthit.chpl.domain.CertificationBody;
 import gov.healthit.chpl.domain.CriteriaSpecificDescriptiveModel;
 import gov.healthit.chpl.domain.DescriptiveModel;
@@ -50,6 +44,7 @@ import lombok.extern.log4j.Log4j2;
 @RestController
 @Loggable
 @Log4j2
+@Deprecated
 public class SearchViewController {
     private static final int MAX_PAGE_SIZE = 100;
 
@@ -73,99 +68,6 @@ public class SearchViewController {
         this.certifiedProductSearchManager = certifiedProductSearchManager;
         this.ff4j = ff4j;
         this.fileUtils = fileUtils;
-    }
-
-    /**
-     * Streams a file back to the end user for the specified edition (2011,2014,2015,etc)
-     * in the specified format (xml or csv). Optionally will send back the definition
-     * files instead. The file that is sent back is generated via a quartz job on a
-     * regular basis.
-     * @param editionInput 2011, 2014, or 2015
-     * @param formatInput csv or xml
-     * @param isDefinition whether to send back the data file or definition file
-     * @param request http request
-     * @param response http response, used to stream back the file
-     * @throws IOException if the file cannot be read
-     */
-    @ApiOperation(value = "Download the entire CHPL as XML.",
-            notes = "Once per day, the entire certified product listing is "
-                    + "written out to XML files on the CHPL servers, one for each "
-                    + "certification edition. This method allows any user to download "
-                    + "that XML file. It is formatted in such a way that users may import "
-                    + "it into Microsoft Excel or any other XML tool of their choosing. To download "
-                    + "any one of the XML files, append ‘&edition=year’ to the end of the query string "
-                    + "(e.g., &edition=2015). A separate query is required to download each of the XML files.")
-    @RequestMapping(value = "/download", method = RequestMethod.GET, produces = "application/xml")
-    public void download(@RequestParam(value = "edition", required = false) String editionInput,
-            @RequestParam(value = "format", defaultValue = "xml", required = false) String formatInput,
-            @RequestParam(value = "definition", defaultValue = "false", required = false) Boolean isDefinition,
-            HttpServletRequest request, HttpServletResponse response) throws IOException {
-        //parse inputs
-        String edition = editionInput;
-        String format = formatInput;
-        String responseType = "text/csv";
-        String filenameToStream = null;
-
-        if (!StringUtils.isEmpty(edition)) {
-            // make sure it's a 4 character year
-            edition = edition.trim();
-            if (!edition.startsWith("20")) {
-                edition = "20" + edition;
-            }
-        } else {
-            edition = "all";
-        }
-
-        if (!StringUtils.isEmpty(format) && format.equalsIgnoreCase("csv")) {
-            format = "csv";
-        } else {
-            format = "xml";
-            responseType = "application/xml";
-        }
-
-        File toDownload = null;
-        //if the user wants a definition file, find it
-        if (isDefinition != null && isDefinition.booleanValue()) {
-            if (format.equals("xml")) {
-                toDownload = fileUtils.getDownloadFile(env.getProperty("schemaXmlName"));
-            } else if (edition.equals("2014")) {
-                toDownload = fileUtils.getDownloadFile(env.getProperty("schemaCsv2014Name"));
-            } else if (edition.equals("2015")) {
-                if (ff4j.check(FeatureList.RWT_ENABLED)) {
-                    toDownload = fileUtils.getDownloadFile(env.getProperty("schemaCsv2015NameWithRWT"));
-                } else {
-                    toDownload = fileUtils.getDownloadFile(env.getProperty("schemaCsv2015Name"));
-                }
-                filenameToStream = env.getProperty("schemaCsv2015Name");
-            }
-
-            if (!toDownload.exists()) {
-                response.getWriter()
-                .write(String.format(messageSource.getMessage(
-                        new DefaultMessageSourceResolvable("resources.schemaFileNotFound"),
-                        LocaleContextHolder.getLocale()), toDownload.getAbsolutePath()));
-                return;
-            }
-        } else {
-            File newestFileWithFormat = fileUtils.getNewestFileMatchingName("^chpl-" + edition + "-.+\\." + format + "$");
-            if (newestFileWithFormat != null) {
-                toDownload = newestFileWithFormat;
-            } else {
-                response.getWriter()
-                .write(String.format(messageSource.getMessage(
-                        new DefaultMessageSourceResolvable(
-                                "resources.fileWithEditionAndFormatNotFound"),
-                        LocaleContextHolder.getLocale()), edition, format));
-                return;
-            }
-        }
-
-        LOGGER.info("Downloading " + toDownload.getName());
-        if (filenameToStream != null) {
-            fileUtils.streamFileAsResponse(toDownload, responseType, response, filenameToStream);
-        } else {
-            fileUtils.streamFileAsResponse(toDownload, responseType, response);
-        }
     }
 
     /**
